requirements.txt

robotframework
robotframework-requests==0.9.2
requests
grequests
bs4
mock
pydantic
